+++ 数据结构/链表的学习/简易通讯录(链表).h
@@ -4,7 +4,7 @@
 
 typedef struct student
 {
-    int id;
+    unsigned long int id;
 
     struct bir
     {
